Split the character processing from the interrupt handler using ring buffer
