+++ kernel/src/gui/mode.h
@@ -1,7 +1,5 @@
 #ifndef MODE_H
 #define MODE_H
 #pragma once
-
-
-
+extern volatile uint64_t mode;
 #endif
